requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.3.2
openai==0.28.0
packaging==25.0
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
safetensors==0.5.3
sniffio==1.3.1
SQLAlchemy==2.0.42
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
